[SPARK-55910][SQL][TESTS] Merge SQLTestUtils into QueryTest
Move all non-SparkFunSuite-dependent methods from `SQLTestUtils`/`SQLTestUtilsBase`
into `QueryTestBase`, and SparkFunSuite-dependent methods into `QueryTest`.
`SQLTestUtils` is now a thin alias (`SparkFunSuite with SQLTestUtilsBase with PlanTest`)
with an empty body.

Changes:
- `QueryTestBase` gains: `withTable`, `withView`, `withSQLConf`, `checkAnswer`,
  `testImplicits`, `waitForTasksToFinish`, `setupTestData`, `sql`, etc.
  Also extends `SQLTestData`, `ClassicConversions`, `Eventually`, `BeforeAndAfterAll`.
- `QueryTest` gains: `withTempDir`, `testQuietly`, `testWithWholeStageCodegenOnAndOff`,
  `test` override, `testWithUninterruptibleThread`, `withResourceTempPath`.
- `QueryTestBase` uses general `sql.SparkSession` (not `classic.SparkSession`).
  `sql` lazy val returns `classic.DataFrame` for compatibility.
- `SQLTestData` visibility changed from `private[sql]` to public.
- `SparkPlanTest` now extends `QueryTest` instead of `SparkFunSuite`.
- Minor import fixes in 8 files (unused imports, `ClassicConversions` now inherited).
