+++ sql/core/src/test/scala/org/apache/spark/sql/execution/SparkPlanTest.scala
@@ -19,8 +19,7 @@ package org.apache.spark.sql.execution
 
 import scala.util.control.NonFatal
 
-import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.{classic, DataFrame, Row, SparkSessionProvider, SQLContext}
+import org.apache.spark.sql.{classic, DataFrame, QueryTest, Row, SQLContext}
 import org.apache.spark.sql.catalyst.analysis.UnresolvedAttribute
 import org.apache.spark.sql.catalyst.plans.logical.LocalRelation
 import org.apache.spark.sql.classic.ClassicConversions._
@@ -30,7 +29,7 @@ import org.apache.spark.sql.test.SQLTestUtils
  * Base class for writing tests for individual physical operators. For an example of how this
  * class's test helper methods can be used, see [[SortSuite]].
  */
-private[sql] abstract class SparkPlanTest extends SparkFunSuite with SparkSessionProvider {
+private[sql] abstract class SparkPlanTest extends QueryTest {
   override protected def spark: classic.SparkSession
 
   /**
